reformas finales, modificacion para que cando el usuario ingrese los numeros se modifiquen en los ejex x e y

+++ TrabajoPractico1/lopes.h
@@ -73,7 +73,7 @@ int restarNumeros(int primerNumero, int segundoNumero);
      *
      */
 
-     int factorialFunc(int numero);
+    long long int factorialFunc(int numero);
 
 
     /** \brief Menu de Opciones
@@ -85,6 +85,8 @@ int restarNumeros(int primerNumero, int segundoNumero);
      */
 
      int menu();
+
+
